% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFARI-package.R
\docType{package}
\name{SAFARI-package}
\alias{SAFARI-package}
\title{Shape Analysis for AI-Reconstructed Images}
\description{
The R package \pkg{SAFARI} provides functions for image processing and
shape analysis of regions of interest (ROI) in AI-reconstructed images.
}
\details{
The main functions are summarized as follows:
\itemize{
\item \code{read.image}: loads and preprocesses reconstructed binary images for segmentation;
\item \code{binary.segmentation}: ROI segmentation and optional feature extraction;
\item \code{compute.features}: feature extraction for an individual ROI;
\item \code{rc.plot}: visualize reconstructed binary or segmented images.
}

See the package vignettes for more introduction and demonstration.
}
